requirements.txt
airflow-providers-clickhouse
numpy
pandas
psycopg2-binary
requests
sqlalchemy